Add files via upload
Interactive Figure: On the left until Colombia, countries of origin of surveyors are indicated, while on the right, countries where they earned their higher educational degrees are shown. The connections between countries represent the destinations surveyors migrated to in pursuit of their degrees, with line thickness indicating the number of surveyors in each connection. There is a significant migration of researchers in ecology and evolutionary biology to countries such as the United States.